Redistribute the Course Components Spec to the controller/component host specs.
This would make the specs test exactly one thing. The Course Components Spec was straddling between controller and component host specs.
